refactored and fixed a bug in characteristic read
